+++ backend2/src/main/java/com/cmoney/backend2/base/model/log/ApiLog.kt
@@ -1,6 +1,7 @@
 package com.cmoney.backend2.base.model.log
 
 import com.google.gson.annotations.SerializedName
+import java.net.URLEncoder
 
 /**
  * cmoneyapi-trace-context的紀錄，每隻都需要使用。
@@ -12,7 +13,7 @@ import com.google.gson.annotations.SerializedName
  * @property model 手機型號
  * @property osVersion 系統版本
  */
-data class ApiLog(
+class ApiLog private constructor(
     @SerializedName("appId")
     val appId : Int,
     @SerializedName("platform")
@@ -25,4 +26,36 @@ data class ApiLog(
     val model: String,
     @SerializedName("osVersion")
     val osVersion: String
-)
+) {
+    companion object {
+
+        /**
+         * 創建ApiLog
+         *
+         * @param appId App的Id
+         * @param platform 平台
+         * @param appVersion App版本
+         * @param manufacturer 製造商
+         * @param model 手機型號
+         * @param osVersion 系統版本
+         * @return
+         */
+        fun create(
+            appId: Int,
+            platform: Int,
+            appVersion: String,
+            manufacturer: String,
+            model: String,
+            osVersion: String
+        ): ApiLog {
+            return ApiLog(
+                appId = appId,
+                platform = platform,
+                appVersion = URLEncoder.encode(appVersion, Charsets.UTF_8.name()),
+                manufacturer = URLEncoder.encode(manufacturer, Charsets.UTF_8.name()),
+                model = URLEncoder.encode(model, Charsets.UTF_8.name()),
+                osVersion = URLEncoder.encode(osVersion, Charsets.UTF_8.name())
+            )
+        }
+    }
+}