+++ backend2/src/main/java/com/cmoney/backend2/customgroup2/service/api/searchstocks/SearchStocksRequestBody.kt
@@ -4,7 +4,5 @@ import com.google.gson.annotations.SerializedName
 
 data class SearchStocksRequestBody(
     @SerializedName("keyword")
-    val keyword: String,
-    @SerializedName("language")
-    val language: String
+    val keyword: String
 )