+++ backend2/src/main/java/com/cmoney/backend2/customgroup2/service/api/searchstocksbymarkettype/SearchStocksByMarketTypeRequestBody.kt
@@ -1,12 +1,11 @@
 package com.cmoney.backend2.customgroup2.service.api.searchstocksbymarkettype
 
+import com.cmoney.backend2.customgroup2.service.api.data.RequestMarketType
 import com.google.gson.annotations.SerializedName
 
 data class SearchStocksByMarketTypeRequestBody(
     @SerializedName("keyword")
     val keyword: String,
-    @SerializedName("language")
-    val language: String,
-    @SerializedName("marketTypes")
-    val marketTypes: List<String>
+    @SerializedName("commodityTypes")
+    val marketTypes: List<RequestMarketType>
 )