+++ backend2/src/main/java/com/cmoney/backend2/userbehavior/service/api/common/Event.kt
@@ -1,0 +1,12 @@
+package com.cmoney.backend2.userbehavior.service.api.common
+
+import com.google.gson.annotations.SerializedName
+
+data class Event(
+    @SerializedName("info")
+    val descriptions: Map<String, Any>?,
+    @SerializedName("name")
+    val name: String,
+    @SerializedName("time")
+    val time: Long
+)