+++ CHANGELOG.md
@@ -1,41 +1,167 @@
-#  (2021-08-12)
+#  (2021-09-23)
+
+
+
+# [4.7.0](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/compare/4.6.0...4.7.0) (2021-09-23)
+
+
+### Bug Fixes
+
+* 修正ApiLog欄位 Encode的方式 ([22377dc](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/commits/22377dcbe8fa5fb70c18dccbe7db6c4b0ffc1afa))
+* 修正Header ApiLog字串經過URL Encode ([b413e18](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/commits/b413e18da19c1097ccf946299c69603304054fe1))
+
+
+### Features
+
+* 圖片上傳服務新增GenreAndSubGenre參數物件，新增401測試，修正其餘測試 ([ad4bee3](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/commits/ad4bee323f6774e8f592bbcacc3be8196bb74537))
+
+
+
+#  (2021-09-08)
+
+
+
+# [4.6.0](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/compare/4.5.1...4.6.0) (2021-09-08)
+
+
+### Features
+
+* 新增 Note 服務回應的 tag 欄位([fa1835bb](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/commit/fa1835bbe92a3d7354e251c9f3e8b6ed186fcfaa))
+
+
+#  (2021-09-04)
+
+
+
+## [4.5.1](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/compare/4.5.0...4.5.1) (2021-09-04)
+
+
+### Bug Fixes
+
+* 修正UserBehavior沒加入Proguard ([0239681](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/commits/0239681ef32fd4c22863e3251067f60a538523fa))
+
+
+
+# [4.5.0](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/compare/4.4.0...4.5.0) (2021-09-04)
+
+
+### Features
+
+* 新增使用者分析API ([45661cc](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/commits/45661ccaab4eee60f2fc6682ddf848885ed41551))
+
+
+
+# [4.4.0](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/compare/4.4.0-beta01...4.4.0) (2021-08-23)
+
+
+### Bug Fixes
+
+* 修正MobileService-追訊，服務3-4. 我也想知道回答 action沒有全小寫 ([c224f06](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/commits/c224f06391edc99911b76a52416ab3c8d5555396))
+
+
+### Features
+
+* 新增判斷用戶是否有CMoney續約中手機商品 ([03d4d4a](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/commits/03d4d4a24dffd7fe58a443aba90ec416b290186f))
+* 補RecordApi ([4f12205](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/commits/4f12205b3b9da7cc3f34cb989669fabdef269b04))
+
+
+
+# [4.4.0-beta01](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/compare/4.3.0...4.4.0-beta01) (2021-08-20)
+
+
+### Features
+
+* 新增權限Type列舉 ([edfe47b](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/commits/edfe47b85448ac68af50f6057904667ae13e8810))
+* 新增語系Header物件，新增自選股服務語系參數 ([9ce9868](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/commits/9ce986842d1b1f307d1ecd3054ddf0065450213c))
 
 
 
 # [4.3.0](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/compare/4.3.0-alpha07...4.3.0) (2021-08-12)
 
 
+### Features
+
+* 新增ChipKWeb 服務6-9. 要求國際盤後資料 ([88581b4](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/commits/88581b44e8163e8be54f26b9860a7ddb926d2864))
+
+
+
+# [4.3.0-alpha07](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/compare/4.3.0-alpha06...4.3.0-alpha07) (2021-08-11)
+
+
+### Features
+
+* 新增可取得他人的自我介紹欄位 ([d0f32e7](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/commits/d0f32e7ff65e72d88ce749abdc0ebf727a715354))
+
+
+
+# [4.3.0-alpha06](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/compare/4.3.0-alpha05...4.3.0-alpha06) (2021-08-10)
+
+
+### Features
+
+* 調整附加資訊Service可以呼叫不同的domain ([b0f09fa](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/commits/b0f09fac21a1d05f2d0d6d16fa588805bfa1934f))
+
+
+
+# [4.3.0-alpha05](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/compare/4.3.0-alpha04...4.3.0-alpha05) (2021-08-06)
+
+
+### Bug Fixes
+
+* 修正文章狀態欄位放錯位置問題 ([04affbc](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/commits/04affbcac3c2d4d6b88d53f2b8813a5c2a0e5762))
+
+
+
+# [4.3.0-alpha04](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/compare/4.3.0-alpha03...4.3.0-alpha04) (2021-08-05)
+
+
+### Bug Fixes
+
+* 修正發文時股票Tag的猜多空狀態使用String傳出 應該要以Int形式傳出 ([155ee80](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/commits/155ee803cf3de2ea3abc0da2af3b64a9fd9c3e25))
+
+
+
+# [4.3.0-alpha03](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/compare/4.3.0-alpha02...4.3.0-alpha03) (2021-08-05)
+
+
 ### Bug Fixes
 
 * createComment createTime ms -> s ([635b097](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/commits/635b09795c52700a05e35317d252cbf3ad5ce522))
-* 修正發文時股票Tag的猜多空狀態使用String傳出 應該要以Int形式傳出 ([155ee80](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/commits/155ee803cf3de2ea3abc0da2af3b64a9fd9c3e25))
-* 修正文章狀態欄位放錯位置問題 ([04affbc](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/commits/04affbcac3c2d4d6b88d53f2b8813a5c2a0e5762))
 
 
 ### Features
 
 * 新增Google登入 ([ca1127f](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/commits/ca1127fd52370f5b5c9c3cd0bc824094feef3568))
-* 調整附加資訊Service可以呼叫不同的domain ([b0f09fa](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/commits/b0f09fac21a1d05f2d0d6d16fa588805bfa1934f))
-* 新增可取得他人的自我介紹欄位 ([d0f32e7](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/commits/d0f32e7ff65e72d88ce749abdc0ebf727a715354))
-* 新增ChipKWeb 服務6-9. 要求國際盤後資料 ([88581b4](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/commits/88581b44e8163e8be54f26b9860a7ddb926d2864))
 
 
 
-#  (2021-07-30)
+# [4.3.0-alpha02](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/compare/4.3.0-alpha01...4.3.0-alpha02) (2021-08-03)
+
+
+
+# [4.3.0-alpha01](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/compare/4.2.0...4.3.0-alpha01) (2021-07-30)
 
 
 
 # [4.2.0](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/compare/4.2.0-alpha01...4.2.0) (2021-07-30)
 
 
-### Bug Fixes
 
-* 修正發回文時 。回文的附加多媒體的Key ([fe2b95e](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/commits/fe2b95e72ce87b841ee927339de64610549a3ff4))
+# [4.2.0-alpha01](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/compare/4.1.2-alpha01...4.2.0-alpha01) (2021-07-24)
 
 
 ### Features
 
 * 新增CustomGroup2 service及UnitTest ([9bbc1ac](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/commits/9bbc1aca0a58e9fdbdb9d24ae773dd3d6edc0637))
+
+
+
+## [4.1.2-alpha01](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/compare/4.1.1...4.1.2-alpha01) (2021-07-23)
+
+
+### Bug Fixes
+
+* 修正發回文時 。回文的附加多媒體的Key ([fe2b95e](http://192.168.10.147:10080/CG_Mobile/CG_Module_Android/Backend2/backend2/commits/fe2b95e72ce87b841ee927339de64610549a3ff4))